+++ src/editor/components/iomd-editor/__tests__/fetch-line-suggestion.test.js
@@ -3,6 +3,7 @@ import { fetchLineSuggestion } from "../fetch-line-suggestion";
 const nonFileNameSuggestionLabels = [
   "arrayBuffer",
   "blob",
+  "bytes",
   "css",
   "javascript",
   "js",