+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Report more specific errors in console when executing fetch cell (fixes #1853) (#2475)
 - Better onboarding around the use of the "run all" button (#2471)
 - Automatically prune revision history (fixes #2451) (#2461)
+- Add `bytes` fetch to fetch chunks (see docs) (fixes #1786)
 
 # 0.15.1 (2019-11-06)